requirements.txt
cryptography==39.0.1
cryptography==41.0.3
gTTS==2.3.2
python_vlc==3.0.18122
